feat(dashboard): hybrid agents empty state with ff-gated CTA
