+++ apps/dashboard/src/pages/agents.tsx
@@ -1,19 +1,9 @@
 import { FeatureFlagsKeysEnum } from '@novu/shared';
 import { CaretSortIcon } from '@radix-ui/react-icons';
 import { useMutation } from '@tanstack/react-query';
-import type { FormEvent, ReactElement, ReactNode } from 'react';
+import type { FormEvent, ReactElement } from 'react';
 import { useCallback, useId, useMemo, useState } from 'react';
-import {
-  RiArrowRightSLine,
-  RiChat3Line,
-  RiCheckLine,
-  RiCloseLine,
-  RiGithubFill,
-  RiMailLine,
-  RiMessage3Line,
-  RiMoreLine,
-  RiUser3Line,
-} from 'react-icons/ri';
+import { RiArrowRightSLine, RiCheckLine, RiCloseLine, RiMailLine, RiMessage3Line, RiMoreLine } from 'react-icons/ri';
 import {
   SiGithub,
   SiGooglechat,
@@ -25,6 +15,7 @@ import {
   SiZoom,
 } from 'react-icons/si';
 import { NovuApiError, post } from '@/api/api.client';
+import { AgentsEmptyTeaser } from '@/components/agents/agents-empty-teaser';
 import { AgentsList } from '@/components/agents/agents-list';
 import { DashboardLayout } from '@/components/dashboard-layout';
 import { PageMeta } from '@/components/page-meta';
@@ -41,7 +32,6 @@ import { useFeatureFlag } from '@/hooks/use-feature-flag';
 import { cn } from '@/utils/ui';
 
 const slackIcon = '/images/providers/light/square/slack.svg';
-const msTeamsIcon = '/images/providers/light/square/msteams.svg';
 const discordIcon = '/images/providers/light/square/discord.svg';
 
 const AGENT_RUN_OPTIONS = [
@@ -113,24 +103,6 @@ const PROVIDER_DEFINITIONS: ProviderDefinition[] = [
   },
   { id: 'other', label: 'Other', icon: <RiMoreLine className="size-4 shrink-0 text-text-sub" aria-hidden /> },
 ];
-
-type AgentsPillProps = {
-  children: ReactNode;
-  className?: string;
-};
-
-function AgentsPill({ children, className }: AgentsPillProps) {
-  return (
-    <span
-      className={cn(
-        'border-stroke-soft bg-bg-weak text-text-strong inline-flex items-center gap-1 rounded border px-1 py-0.5 text-label-sm font-medium',
-        className
-      )}
-    >
-      {children}
-    </span>
-  );
-}
 
 type AgentsEarlyAccessDialogProps = {
   open: boolean;
@@ -440,100 +412,20 @@ export function AgentsPage() {
         {isConversationalAgentsEnabled ? (
           <AgentsList />
         ) : (
-          <div className="flex min-h-[min(720px,calc(100vh-8rem))] flex-col items-center justify-center px-4 py-10 md:px-8">
-            <div className="flex w-full max-w-[1133px] flex-col items-stretch gap-12">
-              <img
-                src="/images/agents-teaser.svg"
-                alt=""
-                className="block h-auto w-full max-w-[456px]"
-                width={456}
-                height={256}
-              />
-
-              <div className="flex w-full max-w-[700px] flex-col items-start gap-3 self-start">
-                <div className="flex flex-col gap-1 text-left">
-                  <p className="text-text-strong text-[16px] font-medium leading-6 tracking-[-0.176px]">
-                    Unified conversational API for AI agents
-                  </p>
-                  <p className="text-text-soft text-[14px] font-medium leading-5 tracking-[-0.084px]">
-                    You own the agent Brain, and Novu gives it voice. Distribute your agent across multiple channels
-                    with a unified API.
-                  </p>
-                </div>
-
-                <ul className="flex flex-col gap-1.5 py-3">
-                  <li className="text-text-sub flex flex-wrap items-center gap-1 text-[14px] font-medium leading-5 tracking-[-0.084px]">
-                    <RiCheckLine className="text-success size-3 shrink-0" aria-hidden />
-                    <span>Connect your agent to</span>
-                    <AgentsPill className="-rotate-1">
-                      <img src={slackIcon} alt="" className="size-3.5" />
-                      <span>Slack</span>
-                    </AgentsPill>
-                    <AgentsPill className="rotate-1">
-                      <RiGithubFill className="size-3.5 shrink-0" aria-hidden />
-                      <span>GitHub</span>
-                    </AgentsPill>
-                    <AgentsPill className="-rotate-1">
-                      <img src={msTeamsIcon} alt="" className="size-3.5" />
-                      <span>MS Teams</span>
-                    </AgentsPill>
-                    <AgentsPill className="rotate-1">
-                      <SiWhatsapp className="size-3.5 shrink-0 text-[#25D366]" aria-hidden />
-                      <span>WhatsApp</span>
-                    </AgentsPill>
-                    <AgentsPill className="-rotate-1">
-                      <SiLinear className="text-text-strong size-3.5 shrink-0" aria-hidden />
-                      <span>Linear</span>
-                    </AgentsPill>
-                    <span>and +15 more.</span>
-                  </li>
-
-                  <li className="text-text-sub flex flex-wrap items-center gap-1 text-[14px] font-medium leading-5 tracking-[-0.084px]">
-                    <RiCheckLine className="text-success size-3 shrink-0" aria-hidden />
-                    <span>Access conversation history</span>
-                    <AgentsPill>
-                      <RiChat3Line className="text-text-sub size-3" aria-hidden />
-                      <span className="font-mono text-[14px] tracking-[-0.28px]">5</span>
-                    </AgentsPill>
-                    <span>
-                      and state via unified <code className="text-text-strong font-mono text-[14px]">agent()</code>{' '}
-                      handler.
-                    </span>
-                  </li>
-
-                  <li className="text-text-sub flex flex-wrap items-center gap-1 text-[14px] font-medium leading-5 tracking-[-0.084px]">
-                    <RiCheckLine className="text-success size-3 shrink-0" aria-hidden />
-                    <span>Provider identities resolved to</span>
-                    <AgentsPill>
-                      <span className="bg-neutral-alpha-200 flex size-4 items-center justify-center rounded-full">
-                        <RiUser3Line className="text-text-sub size-3" aria-hidden />
-                      </span>
-                      <span className="text-text-strong">Subscriber</span>
-                    </AgentsPill>
-                    <span>entity.</span>
-                  </li>
-
-                  <li className="text-text-sub flex flex-wrap items-center gap-1 text-[14px] font-medium leading-5 tracking-[-0.084px]">
-                    <RiCheckLine className="text-success size-3 shrink-0" aria-hidden />
-                    <span>Rich provider interactions, like action buttons, attachments, reactions, and more.</span>
-                  </li>
-                </ul>
-
-                <div className="flex w-full justify-start">
-                  <Button
-                    variant="secondary"
-                    mode="gradient"
-                    size="xs"
-                    trailingIcon={RiArrowRightSLine}
-                    type="button"
-                    onClick={() => setEarlyAccessOpen(true)}
-                  >
-                    Request early access
-                  </Button>
-                </div>
-              </div>
-            </div>
-          </div>
+          <AgentsEmptyTeaser
+            cta={
+              <Button
+                variant="secondary"
+                mode="gradient"
+                size="xs"
+                trailingIcon={RiArrowRightSLine}
+                type="button"
+                onClick={() => setEarlyAccessOpen(true)}
+              >
+                Request early access
+              </Button>
+            }
+          />
         )}
       </DashboardLayout>
     </>
